⭐ feat(maitred): Update maitred - hookup to the API
