+++ infra/realtime.ts
@@ -1,0 +1,3 @@
+export const device = new sst.aws.Realtime("Realtime", {
+    authorizer: "./packages/functions/src/realtime/authorizer.handler"
+})
