+++ packages/maitred/internal/auth/auth.go
@@ -4,13 +4,9 @@ import (
 	"encoding/json"
 	"fmt"
 	"io"
-	"nestri/maitred/pkg/resource"
+	"nestri/maitred/internal/resource"
 	"net/http"
 	"net/url"
-	"os"
-	"os/exec"
-
-	"github.com/charmbracelet/log"
 )
 
 type UserCredentials struct {
@@ -18,18 +14,12 @@ type UserCredentials struct {
 	RefreshToken string `json:"refresh_token"`
 }
 
-func FetchUserToken(teamSlug string) (*UserCredentials, error) {
-	hostname, err := os.Hostname()
-	if err != nil {
-		log.Fatal("Could not get the hostname")
-	}
+func FetchUserToken() (*UserCredentials, error) {
 	data := url.Values{}
 	data.Set("grant_type", "client_credentials")
-	data.Set("client_id", "device")
+	data.Set("client_id", "machine")
 	data.Set("client_secret", resource.Resource.AuthFingerprintKey.Value)
-	data.Set("team", teamSlug)
-	data.Set("hostname", hostname)
-	data.Set("provider", "device")
+	data.Set("provider", "machine")
 	resp, err := http.PostForm(resource.Resource.Auth.Url+"/token", data)
 	if err != nil {
 		return nil, err
@@ -46,13 +36,4 @@ func FetchUserToken(teamSlug string) (*UserCredentials, error) {
 		return nil, err
 	}
 	return &credentials, nil
-}
-
-func GetHostname() string {
-	cmd, err := exec.Command("cat", "/etc/hostname").Output()
-	if err != nil {
-		log.Error("error getting container hostname", "err", err)
-	}
-	output := string(cmd)
-	return output
 }
