+++ packages/maitred/internal/realtime/utils.go
@@ -1,11 +1,11 @@
-package party
+package realtime
 
 import (
 	"fmt"
 	"os"
 	"time"
 
-	"math/rand"
+	"crypto/rand"
 
 	"github.com/charmbracelet/log"
 	"github.com/oklog/ulid/v2"
@@ -20,12 +20,12 @@ var (
 )
 
 func generateClientID() string {
-	// Create a source of entropy (use cryptographically secure randomness in production)
-	entropy := rand.New(rand.NewSource(time.Now().UnixNano()))
+	// Create a source of entropy (cryptographically secure)
+	entropy := ulid.Monotonic(rand.Reader, 0)
 
 	// Generate a new ULID
 	id := ulid.MustNew(ulid.Timestamp(time.Now()), entropy)
 
 	// Create the client ID string
-	return fmt.Sprintf("client_%s", id.String())
+	return fmt.Sprintf("mch_%s", id.String())
 }
