+++ packages/maitred/main.go
@@ -1,21 +1,21 @@
 package main
 
 import (
-	"nestri/maitred/pkg/party"
+	"nestri/maitred/internal/realtime"
 	"os"
 
 	"github.com/charmbracelet/log"
 )
 
 func main() {
-	var teamSlug string //FIXME: Switch to team-slug as they are more memorable but still unique
+	var secretToken string //FIXME: Switch to team-slug as they are more memorable but still unique
 
 	if len(os.Args) > 1 {
-		teamSlug = os.Args[1]
+		secretToken = os.Args[1]
 	} else {
 		log.Fatal("Nestri needs a team slug to register this container to")
 	}
-	party.Run(teamSlug)
+	realtime.Run(secretToken)
 
 	//TODO: On stop here, set the API as the instance is not running (stopped)
 }
